Link some pages and bugfix
Some rare visual bug caused every div inside the body left a space to the right. Happened because I'm using
width:100vw, and in top of that I'm putting padding to some divs, so they expand to much and so the page, but
not every other content. Now should be fixed.